+++ app/src/main/kotlin/com/olderwold/sliide/data/dto/NewUserDTO.kt
@@ -1,12 +1,12 @@
-package com.olderwold.sliide.data
+package com.olderwold.sliide.data.dto
 
 import com.google.gson.annotations.SerializedName
 
-internal data class DeleteUserDto(
+internal data class NewUserDTO(
     @SerializedName("code")
     val code: Int? = null,
     @SerializedName("data")
-    val data: Any? = null,
+    val data: UserDto? = null,
     @SerializedName("meta")
     val meta: Any? = null
 )
