+++ app/src/main/kotlin/com/olderwold/sliide/data/dto/UserMapper.kt
@@ -1,0 +1,26 @@
+package com.olderwold.sliide.data.dto
+
+import com.olderwold.sliide.data.mapper.GenderStatusMapper
+import com.olderwold.sliide.data.mapper.UserStatusMapper
+import com.olderwold.sliide.domain.User
+
+internal class UserMapper(
+    private val genderStatusMapper: GenderStatusMapper,
+    private val userStatusMapper: UserStatusMapper,
+) {
+    @Suppress("ReturnCount")
+    fun map(dto: UserDto?): User? {
+        val data = dto ?: return null
+        val id = data.id ?: return null
+        val gender = genderStatusMapper.map(data.gender)
+        val status = userStatusMapper.map(data.status)
+        return User(
+            id = id,
+            email = data.email,
+            gender = gender,
+            status = status,
+            createdAt = data.createdAt,
+            updatedAt = data.updatedAt,
+        )
+    }
+}
