+++ app/src/main/kotlin/com/olderwold/sliide/data/mapper/UserListMapper.kt
@@ -1,10 +1,12 @@
-package com.olderwold.sliide.data
+package com.olderwold.sliide.data.mapper
 
+import com.olderwold.sliide.data.dto.UserListDTO
+import com.olderwold.sliide.data.dto.UserMapper
 import com.olderwold.sliide.domain.UserList
 
 internal class UserListMapper(
     private val paginationMapper: PaginationMapper,
-    private val userDtoMapper: UserDto.Mapper,
+    private val userDtoMapper: UserMapper,
 ) {
     fun map(dto: UserListDTO): UserList {
         val users = dto.data.filterNotNull()
