+++ app/src/main/kotlin/com/olderwold/sliide/data/usecase/NetworkDeleteUser.kt
@@ -1,6 +1,7 @@
-package com.olderwold.sliide.data
+package com.olderwold.sliide.data.usecase
 
-import com.olderwold.sliide.domain.DeleteUser
+import com.olderwold.sliide.data.GoRestClient
+import com.olderwold.sliide.domain.usecase.DeleteUser
 import com.olderwold.sliide.domain.User
 import com.olderwold.sliide.rx.Schedulers
 import dagger.Binds
