+++ app/src/main/kotlin/com/olderwold/sliide/domain/usecase/CreateUser.kt
@@ -1,5 +1,6 @@
-package com.olderwold.sliide.domain
+package com.olderwold.sliide.domain.usecase
 
+import com.olderwold.sliide.domain.User
 import io.reactivex.Completable
 
 internal interface CreateUser {
