+++ app/src/main/kotlin/com/olderwold/sliide/domain/usecase/GetLatestUserList.kt
@@ -1,5 +1,6 @@
-package com.olderwold.sliide.domain
+package com.olderwold.sliide.domain.usecase
 
+import com.olderwold.sliide.domain.UserList
 import io.reactivex.Single
 
 internal interface GetLatestUserList {
