+++ app/src/main/kotlin/com/olderwold/sliide/presentation/create/CreateUserViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.SingleEvent
 import androidx.lifecycle.toEvent
 import com.olderwold.sliide.R
-import com.olderwold.sliide.domain.CreateUser
+import com.olderwold.sliide.domain.usecase.CreateUser
 import com.olderwold.sliide.domain.User
 import com.olderwold.sliide.rx.RxOperators
 import com.olderwold.sliide.rx.RxViewModel
