+++ app/src/test/kotlin/com/olderwold/sliide/data/usecase/NetworkGetLatestUserListTest.kt
@@ -1,5 +1,6 @@
-package com.olderwold.sliide.data
+package com.olderwold.sliide.data.usecase
 
+import com.olderwold.sliide.data.GoRestClient
 import com.olderwold.sliide.domain.Pagination
 import com.olderwold.sliide.domain.UserList
 import com.olderwold.sliide.rx.ImmediateSchedulers
